refactored tests for minimal code repetition
