+++ src/test/java/model/data/process/functions/StandardDeviationTest.java
@@ -20,66 +20,7 @@ import static org.junit.Assert.*;
  * 
  * @author Louis Gosschalk 12-05-2015
  */
-public class StandardDeviationTest {
-
-	private DataTable table;
-	private DataColumn stringColumn;
-	private DataColumn intColumn;
-	private DataColumn intsColumn;
-	private DataColumn floatColumn;
-	private DataColumn floatsColumn;
-	private DataColumn floatersColumn;
-
-	/**
-	 * simulate datamodel.
-	 * 
-	 * @throws Exception
-	 */
-	@Before
-	public void setUp() throws Exception {
-		DataTableBuilder builder = new DataTableBuilder();
-
-		stringColumn = builder.createColumn("string", StringValue.class);
-		intColumn = builder.createColumn("int", FloatValue.class);
-		intsColumn = builder.createColumn("ints", FloatValue.class);
-		floatColumn = builder.createColumn("float", FloatValue.class);
-		floatsColumn = builder.createColumn("floats", FloatValue.class);
-		floatersColumn = builder.createColumn("floaters", FloatValue.class);
-
-		StringValue string = new StringValue("What");
-		FloatValue int1 = new FloatValue(9);
-		FloatValue int2 = new FloatValue(12);
-		FloatValue float1 = new FloatValue(6.9f);
-		FloatValue float2 = new FloatValue(8.8f);
-		FloatValue float3 = new FloatValue(6.6f);
-		builder.createRow(string, int1, int2, float1, float2, float3);
-
-		string = new StringValue("Can");
-		int1 = new FloatValue(5);
-		int2 = new FloatValue(10);
-		float1 = new FloatValue(6.5f);
-		float2 = new FloatValue(6.9f);
-		float3 = new FloatValue(6.6f);
-		builder.createRow(string, int1, int2, float1, float2, float3);
-
-		string = new StringValue("You");
-		int1 = new FloatValue(3);
-		int2 = new FloatValue(3);
-		float1 = new FloatValue(5.9f);
-		float2 = new FloatValue(8.8f);
-		float3 = new FloatValue(6.4f);
-		builder.createRow(string, int1, int2, float1, float2, float3);
-
-		string = new StringValue("Do");
-		int1 = new FloatValue(10);
-		int2 = new FloatValue(12);
-		float1 = new FloatValue(6.2f);
-		float2 = new FloatValue(5.3f);
-		float3 = new FloatValue(6.6f);
-		builder.createRow(string, int1, int2, float1, float2, float3);
-		builder.setName("test");
-		table = builder.build();
-	}
+public class StandardDeviationTest extends FunctionTest {
 
 	/**
 	 * column of strings should throw exception.
