+++ src/test/java/model/data/process/functions/SumTest.java
@@ -22,67 +22,7 @@ import static org.junit.Assert.*;
  * 
  * @author Louis Gosschalk 16-05-2015
  */
-public class SumTest {
-
-	private DataTable table;
-	private DataColumn stringColumn;
-	private DataColumn intColumn;
-	private DataColumn intsColumn;
-	private DataColumn floatColumn;
-	private DataColumn floatsColumn;
-	private DataColumn floatersColumn;
-
-	/**
-	 * simulate datamodel with single sum for each column type.
-	 * 
-	 * @throws Exception
-	 */
-	@Before
-	public void setUp() throws Exception {
-		DataTableBuilder builder = new DataTableBuilder();
-
-		stringColumn = builder.createColumn("string", StringValue.class);
-		intColumn = builder.createColumn("int", IntValue.class);
-		intsColumn = builder.createColumn("ints", IntValue.class);
-		floatColumn = builder.createColumn("float", FloatValue.class);
-		floatsColumn = builder.createColumn("floats", FloatValue.class);
-		floatersColumn = builder.createColumn("floaters", FloatValue.class);
-
-		StringValue string = new StringValue("What");
-		IntValue int1 = new IntValue(9);
-		IntValue int2 = new IntValue(12);
-		FloatValue float1 = new FloatValue(6.9f);
-		FloatValue float2 = new FloatValue(8.8f);
-		FloatValue float3 = new FloatValue(6.6f);
-		builder.createRow(string, int1, int2, float1, float2, float3);
-
-		string = new StringValue("Can");
-		int1 = new IntValue(5);
-		int2 = new IntValue(10);
-		float1 = new FloatValue(6.5f);
-		float2 = new FloatValue(6.9f);
-		float3 = new FloatValue(6.6f);
-		builder.createRow(string, int1, int2, float1, float2, float3);
-
-		string = new StringValue("You");
-		int1 = new IntValue(3);
-		int2 = new IntValue(3);
-		float1 = new FloatValue(5.9f);
-		float2 = new FloatValue(8.8f);
-		float3 = new FloatValue(6.4f);
-		builder.createRow(string, int1, int2, float1, float2, float3);
-
-		string = new StringValue("Do");
-		int1 = new IntValue(10);
-		int2 = new IntValue(12);
-		float1 = new FloatValue(6.2f);
-		float2 = new FloatValue(5.3f);
-		float3 = new FloatValue(6.6f);
-		builder.createRow(string, int1, int2, float1, float2, float3);
-
-		builder.setName("test");
-		table = builder.build();
-	}
+public class SumTest extends FunctionTest {
 
 	/**
 	 * column of strings should throw exception.
